+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## master (unreleased)
+* Adds CircleCI install generator. ([@abhaynikam][])
 
 ## 0.3.0 (September 19th, 2020)
 * Adds Bullet install generator. ([@abhaynikam][])